temporary conversion to UTF-8 from windows 1250
this adds a CLI flag to convert input files from windows 1250 encoding
to UTF-8. There seems to be an issue with square brackets in the
html->md library however which isn't resolved by this, so as a temporary
measure I'm replacing square brackets with parantheses.

I filed an issue, so hopefully it gets fixed.